GH-127 moving the integration test with re-usable `testers` in a new `aio-lib-java-events-test` module
