+++ events_test/src/test/resources/workspace.properties
@@ -1,0 +1,39 @@
+#
+# Copyright 2017 Adobe. All rights reserved.
+# This file is licensed to you under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License. You may obtain a copy
+# of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed under
+# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+# OF ANY KIND, either express or implied. See the License for the specific language
+# governing permissions and limitations under the License.
+#
+
+aio_meta_scopes=aio_meta_scopes_changeMe
+# ims_org_id = your IMS Organization ID as shown in your Adobe Developer Console workspace
+aio_ims_org_id=aio_ims_org_id_changeMe
+# consumer_org_id = your consumer Org Id as shown in your Adobe Developer Console workspace
+aio_consumer_org_id=aio_consumer_org_id_changeMe
+# api_key = your credential API Key (Client ID) as shown in in your Adobe Developer Console workspace
+aio_api_key=aio_api_key_changeMe
+# credential_id = your credential id as shown in your Adobe Developer Console workspace
+aio_credential_id=aio_credential_id_changeMe
+# client_secret = your credential Client secret as shown in in your Adobe Developer Console workspace
+aio_client_secret=aio_client_secret_changeMe
+# technical_account_id = your credential Technical account ID as shown in your Adobe Developer Console workspace
+aio_technical_account_id=aio_technical_account_id_changeMe
+# You then have 3 options to configure the privateKey (associated with the public key set in your Adobe Developer Console workspace):
+# * Option 1: use a pcks8 file (using the `pkcs8_file_path` config)
+# * Option 2: use base 64 encoded pcks8 key (using the `encoded_pkcs8` config)
+# * Option 3: use a keystore (using the `pkcs12_file_path`, `pkcs12_alias` and `pkcs12_password` config)
+aio_encoded_pkcs8=changeMe
+
+######################################
+#  I/O Events url
+######################################
+#aio_api_url=https://api.adobe.io
+#aio_publish_url=https://eventsingress.adobe.io
+
+
+
