+++ pom.xml
@@ -29,6 +29,8 @@
     <module>events_ingress</module>
     <module>events_journal</module>
 
+    <module>events_test</module>
+
     <module>aem</module>
 
   </modules>
